プロジェクト設定と型安全性の向上に関する修正
MagEngine.vcxproj:
- `<AdditionalIncludeDirectories>` の重複ディレクトリを整理。
- `<Link>` セクションに `<EnableCOMDATFolding>` と `<OptimizeReferences>` を追加し、リンク時の最適化を有効化。

Enemy.cpp:
- `Enemy::Initialize` 内で `maxRollAngle_` と `maxPitchAngle_` の計算式を変更。
  - `M_PI` を `static_cast<float>(M_PI)` に変更し、型の安全性を向上。

+++ application/Enemy.cpp
@@ -33,8 +33,8 @@ void Enemy::Initialize(Object3dSetup *object3dSetup, const std::string &modelPat
 	targetRotationEuler_ = {0.0f, 0.0f, 0.0f};
 	acceleration_ = 8.0f;					  // 加速度
 	rotationSmoothing_ = 8.0f;				  // 回転の滑らかさ
-	maxRollAngle_ = 45.0f * (M_PI / 180.0f);  // 最大ロール角（45度）
-	maxPitchAngle_ = 30.0f * (M_PI / 180.0f); // 最大ピッチ角（30度）
+	maxRollAngle_ = 45.0f * (static_cast<float>(M_PI) / 180.0f);  // 最大ロール角（45度）
+	maxPitchAngle_ = 30.0f * (static_cast<float>(M_PI) / 180.0f); // 最大ピッチ角（30度）
 	bankingFactor_ = 1.5f;					  // バンキング係数
 
 	//========================================
